requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web interface
streamlit>=1.28.0

# Data storage
python-dotenv>=1.0.0

# Optional: for enhanced scraping
selenium>=4.15.0

# Machine Learning dependencies
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0
torch>=2.0.0
joblib>=1.3.0

# Advanced statistics
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
